Convert Spaces to Teams

+++ src/hooks/useDb.ts
@@ -1,8 +1,8 @@
-import { UserContext } from '../contexts/UserContext'
 import { useContext } from 'react'
 import { useErrorPromptContext } from './useErrorPromptContext'
 import { addDoc, collection, deleteDoc, doc, updateDoc } from 'firebase/firestore'
 import { db } from '../firebase/config'
+import { UserContext } from '../contexts/UserContext'
 
 export const useDb = (col: string) => {
   const userContext = useContext(UserContext)
@@ -15,6 +15,9 @@ export const useDb = (col: string) => {
   const { user } = userContext
 
   const addDocument = async (doc: any) => {
+    if (col === 'teams' && user?.role !== 'admin') {
+      throw new Error('Only admins can add teams')
+    }
     try {
       const ref = await addDoc(collection(db, col), { ...doc, uid: user?.uid })
       return ref
@@ -24,6 +27,9 @@ export const useDb = (col: string) => {
   }
 
   const removeDocument = async (docID: string) => {
+    if (col === 'teams' && user?.role !== 'admin') {
+      throw new Error('Only admins can remove teams')
+    }
     try {
       await deleteDoc(doc(collection(db, col), docID))
     } catch {
@@ -32,6 +38,9 @@ export const useDb = (col: string) => {
   }
 
   const updateDocument = async (docID: string, changesObj: any) => {
+    if (col === 'teams' && user?.role !== 'admin') {
+      throw new Error('Only admins can update teams')
+    }
     try {
       await updateDoc(doc(collection(db, col), docID), changesObj)
     } catch {
